FIX Check destroy exists on object when exporting

+++ forms/gridfield/GridFieldExportButton.php
@@ -150,7 +150,9 @@ class GridFieldExportButton implements GridField_HTMLProvider, GridField_ActionP
 			$fileData .= implode($separator, $columnData);
 			$fileData .= "\n";
 
-			$item->destroy();
+			if($item->hasMethod('destroy')) {
+				$item->destroy();
+			}
 		}
 
 		return $fileData;
